requirements.txt
# Core framework dependencies
fastapi==0.115.6
gunicorn==22.0.0
pydantic==2.11.4
pydantic-settings==2.9.1
email-validator==2.1.0
loguru==0.7.2

# LangChain and AI dependencies
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-mcp-adapters==0.1.1
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
openai==1.79.0

# Database and storage dependencies
pymongo==4.6.0
redis==5.0.1
sqlalchemy==2.0.41
psycopg2-binary==2.9.9
pyodbc==5.0.1
scylla-driver==3.29.3
mem0ai>=0.1.48
qdrant-client==1.12.1

# env
python-dotenv==1.1.0

# Authentication and security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.20
python-dotenv==1.1.0
slowapi==0.1.9

# Async and web dependencies
aiofiles==23.2.1
websockets==13.0.0

# Configuration and logging
pyyaml==6.0.2
structlog==23.2.0
prometheus-client==0.19.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-semantic-conventions==0.44b0
python-json-logger==2.0.7

# Task processing
celery==5.3.4
grpcio-tools==1.62.2

# Web scraping and search
tavily-python==0.7.3
beautifulsoup4==4.12.2

# Document processing
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
markdown-it-py==3.0.0
markdownify==1.1.0
markupsafe==3.0.2
marshmallow==3.26.1
MarkItDown[all]==0.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.28.1
pytest-cov==4.1.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Load testing
locust==2.17.0

# Data analysis and visualization
pandas==2.2.3
protobuf==4.21.6
pydantic-core==2.33.2
# pygraphviz==1.14

# Miscellaneous
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.7
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
ujson==5.10.0
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0

# Web UI
gradio==5.32.1

# add-ons
aiofiles>=23.0.0
backoff>=2.2.0
redis>=4.5.0
passlib